+++ attributes/xdebug.rb
@@ -1,7 +1,7 @@
 # Xdebug attributes
 # Only relevant if the `project.install_dev_tools` attribute is set
 
-default['php']['xdebug']['version']         = '2.3.3'
+default['php']['xdebug']['version']         = '2.6.1'
 
 # Used for provisioning the xdebug CLI wrapper script
 default['php']['xdebug']['idekey']          = 'PHPSTORM'